requirements.txt
typer>=0.12.0
pyyaml>=6.0
jinja2>=3.1.0
rich>=13.0.0
click>=8.0.0